backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
groq==0.11.0
langdetect==1.0.9
transformers==4.44.2
sentencepiece==0.2.0
torch==2.4.1
nltk==3.9.1
reportlab==4.2.5
python-multipart==0.0.9
pydantic==2.9.2
python-dotenv==1.0.1
pypdf==4.3.1
httpx==0.27.2
